% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aachen_color_palette.R
\name{aachen_color}
\alias{aachen_color}
\title{Queries colors from RWTH color palette}
\usage{
aachen_color(colors)
}
\arguments{
\item{colors}{Query of colors (see @display_aachen_colors() ) for querie names}
}
\value{
HEX code of queried colors
}
\description{
Queries colors from RWTH color palette
}
\examples{
aachen_color("maygreen")
}
