% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AachenColorPalette-package.R
\docType{package}
\name{AachenColorPalette-package}
\alias{AachenColorPalette}
\alias{AachenColorPalette-package}
\title{AachenColorPalette: Palette of colors inspired by RWTH Aachen corporate design}
\description{
A color palette inspired by the corporate design of the RWTH Aachen.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/christianholland/AachenColorPalette}
  \item Report bugs at \url{https://github.com/christianholland/AachenColorPalette/issues}
}

}
\author{
\strong{Maintainer}: Christian H. Holland \email{cholland2408@gmail.com} (\href{https://orcid.org/0000-0002-3060-5786}{ORCID})

}
\keyword{internal}
