% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_palette.R
\name{display_aachen_colors}
\alias{display_aachen_colors}
\title{Creates a plot showing all avaiable colors of the Aachen palette}
\usage{
display_aachen_colors()
}
\description{
Creates a plot showing all avaiable colors of the Aachen palette
}
\examples{
display_aachen_colors()
}
