% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{run_progeny}
\alias{run_progeny}
\title{Title}
\usage{
run_progeny(E, M, gene_name = "gene", value_name = "expression",
  id_name = "sample", permutation = 10000, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
