% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{run_viper}
\alias{run_viper}
\title{Title}
\usage{
run_viper(E, regulon, gene_name = "gene", value_name = "expression",
  id_name = "sample", regulator_name = "tf", ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
