% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwth_color_palette.R
\name{rwth_color}
\alias{rwth_color}
\title{Queries colors from RWTH color palette}
\usage{
rwth_color(colors)
}
\arguments{
\item{colors}{Query of colors (see @display_rwth_colors() ) for querie names}
}
\value{
HEX code of queried colors
}
\description{
Queries colors from RWTH color palette
}
\examples{
rwth_color("maygreen")
rwth_color(c("blue", "blue75", "blue50", "blue25"))
}
