% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_color.R
\name{aachen_color}
\alias{aachen_color}
\title{Queries colors from Aachen color palette}
\usage{
aachen_color(...)
}
\arguments{
\item{...}{Query of colors. Query colors by combining color name in lower
case (check figure generated by
\code{\link[=display_aachen_colors]{display_aachen_colors()}} for
color names) and color intensity. For colors with 100\% intensity the color
name alone is sufficient.}
}
\value{
HEX code of queried colors
}
\description{
Queries colors from Aachen color palette
}
\examples{
aachen_color("maygreen")
aachen_color(c("blue", "blue75", "blue50", "blue25", "blue10"))
}
