% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aachen_color_palette.R
\name{display_aachen_colors}
\alias{display_aachen_colors}
\title{Creates a plot showing all avaiable colors of the RWTH palette}
\usage{
display_aachen_colors()
}
\description{
Creates a plot showing all avaiable colors of the RWTH palette
}
\examples{
display_rwth_colors()
}
